refactor: 💡 Move video mock data to MockData.js
