--- apps/e-learning-app/src/app/components/moduleSidebar/MockData.js
@@ -1,38 +1,0 @@
-/*
-!!! THIS SHOULD BE A TEMPORARY FILE !!!
-
-  Mock data used for filling sidebar content.
-  One entry will be either:
-  * title type entry
-  * subtitle type entry
-  
-  A title type entry will have big and bold text.
-  A subtitle type entry will have normal font and size.
-  Both types will require a path to the page with the filtered
-  lessons.
-
-*/
-
-export const MockData = [
-  {
-    subtitle: 'Fronda in literatura interbelica',
-    path: '',
-  },
-  {
-    subtitle: "Ion Vinea: 'Manifest activist catre tinerime'",
-    path: '',
-  },
-  {
-    subtitle: "Tristan Tzara: 'Manifest despre amorul slab si amorul amar'",
-    path: '',
-  },
-  {
-    title: 'Limba si comunicare',
-    subtitles: [
-      {
-        subtitle: 'Punctuatia si justificarile ei sintactice si stilistice',
-        path: '',
-      },
-    ],
-  },
-];
